+++ lib/vor/lexer.ex
@@ -38,7 +38,8 @@ defmodule Vor.Lexer do
                    forall where and or not in start_timer cancel_timer restart_timer
                    retransmit_last_response extern if else
                    sends send broadcast system connect solve noop
-                   map_get map_put map_has map_delete map_size map_sum map_merge)
+                   map_get map_put map_has map_delete map_size map_sum map_merge
+                   every)
 
     type = if name in keywords, do: :keyword, else: :identifier
     {rest, [{type, {line, col}, String.to_atom(name)}], context}